Backend/requirements.txt
pandas~=2.3.3
numpy~=2.3.5
PyPDF2~=3.0.1
langchain~=1.1.0
langchain-groq~=1.1.0
openai~=2.8.1
groq~=0.37.0
fastapi~=0.123.3
uvicorn~=0.38.0
python-dotenv~=1.2.1
rich~=14.2.0
python-multipart~=0.0.20
json5~=0.12.1